+++ _data/press.yml
@@ -1,0 +1,5 @@
+- media: Jacobin
+  url: https://www.jacobinmag.com/2020/10/voices-from-the-valley-silicon-valley-tech-industry-workers
+  title: Voices From the Valley Is a Snapshot of the Tech Industry, as Told By Workers Themselves
+  byline: "Review of Voices From the Valley: Tech Workers Talk About What They Do — and How They Do It, edited by Ben Tarnoff and Moira Weigel"
+  date: 2020-10-30
