+++ _includes/main.scss
@@ -1,4 +1,5 @@
 @import 'variables';
+@import 'event';
 
 body {
   font-family: Helvetica Neue, HelveticaNeue, Helvetica, Arial, system-ui, -apple-system, Segoe UI, Roboto, Ubuntu, Cantarell, Noto Sans, sans-serif;
