+++ _sass/_event.scss
@@ -1,0 +1,22 @@
+.event-card {
+  @supports (display: grid) {
+    align-items: center;
+    display: grid;
+    grid-template-columns: 2rem 1fr;
+    grid-gap: 0.5rem;
+  }
+
+
+}
+
+.event-card__info-column {
+  grid-column: 2;
+}
+
+.event-card__icon {
+  fill: $grey;
+  font-size: 0.55em;
+  font-weight: bold;
+  grid-column: 1;
+  width: 2rem;
+}
